Extract `CompilationConfig` from `config.py`
