+++ vllm/config/__init__.py
@@ -1,0 +1,14 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+
+from .compilation import CompilationConfig, CompilationLevel, PassConfig
+from .config import *
+from .utils import ConfigType, config
+
+__all__ = [
+    "CompilationConfig",
+    "CompilationLevel",
+    "PassConfig",
+    "config",
+    "ConfigType",
+]
